requirements.txt
numpy==1.26.4
pyside6==6.6.0

pyserial==3.5
pyvisa==1.14.1
pyvisa-py==0.7.2
opencv-python==4.9.0.80
chipwhisperer==5.7.0